requirements.txt
langchain
sentence-transformers
faiss-cpu
transformers
langchain-community